fix: address PRRT_kwDOSJAM6s6RPzO0 — split hosted payload volume helpers
